requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-docx==1.1.2
numpy==1.26.4
python-dotenv==1.0.1